requirements.txt
git+https://github.com/canonical/charmcraft#egg=charmcraft
http://archive.ubuntu.com/ubuntu/pool/main/p/python-apt/python-apt_2.4.0ubuntu1.tar.xz; sys_platform == "linux"
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
craft-cli==2.0.1
craft-parts==1.23.0
craft-providers==1.14.1
craft-store==2.4.0
cryptography==41.0.3
Deprecated==1.2.14
humanize==4.7.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==24.2.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
more-itertools==10.0.0
overrides==7.3.1
packaging==23.1
platformdirs==3.9.1
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.11
pydantic-yaml==0.11.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyRFC3339==1.1
python-dateutil==2.8.2
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
rpds-py==0.9.2
SecretStorage==3.3.3
six==1.16.0
snap-helpers==0.4.0
tabulate==0.9.0
types-Deprecated==1.2.9.3
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==1.26.16
wrapt==1.15.0
zipp==3.16.2
click
